requirements.txt
fastapi[all]
uvicorn==0.30.3
diffusers
torch
accelerate
torchvision
pillow
scipy
python-dotenv